Update logic, move tests
